REFACTOR: use NestJS dependency injection in the request scope to obtain a connection from the connection pool
