+++ src/authorization/repository/entity.repository.ts
@@ -1,12 +1,16 @@
-import { Injectable } from '@nestjs/common';
+import { Inject, Injectable } from '@nestjs/common';
 import { UpdateEntityInput } from 'src/schema/graphql.schema';
 import { DataSource } from 'typeorm';
 import EntityModel from '../entity/entity.entity';
 import { BaseRepository } from './base.repository';
+import { TENANT_CONNECTION } from '../../database/database.constants';
 
 @Injectable()
 export class EntityModelRepository extends BaseRepository<EntityModel> {
-  constructor(private dataSource: DataSource) {
+  constructor(
+    @Inject(TENANT_CONNECTION)
+    private dataSource: DataSource,
+  ) {
     super(EntityModel, dataSource);
   }
