+++ src/authorization/repository/user.repository.ts
@@ -1,13 +1,17 @@
-import { Injectable } from '@nestjs/common';
+import { Inject, Injectable } from '@nestjs/common';
 import { UpdateUserInput } from 'src/schema/graphql.schema';
 import { DataSource, In } from 'typeorm';
 import User from '../entity/user.entity';
 import UserGroup from '../entity/userGroup.entity';
 import { BaseRepository } from './base.repository';
+import { TENANT_CONNECTION } from '../../database/database.constants';
 
 @Injectable()
 export class UserRepository extends BaseRepository<User> {
-  constructor(private dataSource: DataSource) {
+  constructor(
+    @Inject(TENANT_CONNECTION)
+    private dataSource: DataSource,
+  ) {
     super(User, dataSource);
   }
 
@@ -16,7 +20,7 @@ export class UserRepository extends BaseRepository<User> {
   }
 
   async getUserByEmail(email: string) {
-    return (await this.getQueryBuilder('user'))
+    return this.createQueryBuilder('user')
       .where('lower(user.email) = lower(:email)', { email })
       .getOne();
   }
@@ -42,14 +46,14 @@ export class UserRepository extends BaseRepository<User> {
   }
 
   async getUsersByGroupId(groupId: string): Promise<User[]> {
-    return (await this.getQueryBuilder('user'))
+    return this.createQueryBuilder('user')
       .leftJoinAndSelect(UserGroup, 'userGroup', 'userGroup.userId = user.id')
       .where('userGroup.groupId = :groupId', { groupId })
       .getMany();
   }
 
   async getUserCountForGroupId(groupId: string): Promise<number> {
-    return (await this.getQueryBuilder('user'))
+    return this.createQueryBuilder('user')
       .innerJoinAndSelect(UserGroup, 'userGroup', 'userGroup.userId = user.id')
       .where('userGroup.groupId = :groupId', { groupId })
       .getCount();
