+++ src/authorization/service/user.service.ts
@@ -1,5 +1,5 @@
 import { BadRequestException, Inject, Injectable } from '@nestjs/common';
-import { DataSource, EntityManager, SelectQueryBuilder } from 'typeorm';
+import { DataSource, SelectQueryBuilder } from 'typeorm';
 import { UserNotAuthorized } from '../../authentication/exception/userauth.exception';
 import { FilterBuilder } from '../../common/filter.builder';
 import { SearchEntity } from '../../constants/search.entity.enum';
@@ -33,7 +33,7 @@ import SearchService from './search.service';
 import { UserServiceInterface } from './user.service.interface';
 import { UserCacheServiceInterface } from './usercache.service.interface';
 import { ExecutionManager } from '../../util/execution.manager';
-import { getConnection } from '../../util/database.connection';
+import { TENANT_CONNECTION } from '../../database/database.constants';
 
 @Injectable()
 export class UserService implements UserServiceInterface {
@@ -49,6 +49,7 @@ export class UserService implements UserServiceInterface {
     private groupCacheService: GroupCacheServiceInterface,
     @Inject(PermissionCacheServiceInterface)
     private permissionCacheService: PermissionCacheServiceInterface,
+    @Inject(TENANT_CONNECTION)
     private dataSource: DataSource,
     private searchService: SearchService,
     @Inject(RoleCacheServiceInterface)
@@ -75,7 +76,7 @@ export class UserService implements UserServiceInterface {
         );
       }
     };
-    const qb = await this.userRepository.getQueryBuilder('user');
+    const qb = this.userRepository.createQueryBuilder('user');
     if (input?.search) {
       this.searchService.generateSearchTermForEntity(
         qb,
@@ -157,7 +158,7 @@ export class UserService implements UserServiceInterface {
       user.groups.map((group) => ({ userId: id, groupId: group })),
     );
 
-    await (await getConnection()).manager.transaction(async (entityManager) => {
+    await this.dataSource.manager.transaction(async (entityManager) => {
       const userGroupsRepo = entityManager.getRepository(UserGroup);
       await userGroupsRepo.remove(groupsToBeRemovedFromUser);
       await userGroupsRepo.save(userGroups);
@@ -202,7 +203,7 @@ export class UserService implements UserServiceInterface {
       })),
     );
 
-    const userPermissionsUpdated = await (await getConnection()).transaction(
+    const userPermissionsUpdated = await this.dataSource.manager.transaction(
       async (entityManager) => {
         const userPermissionsRepo = entityManager.getRepository(UserPermission);
         await userPermissionsRepo.remove(userPermissionsToBeRemoved);
@@ -228,7 +229,7 @@ export class UserService implements UserServiceInterface {
       throw new UserNotFoundException(id);
     }
 
-    await (await getConnection()).manager.transaction(async (entityManager) => {
+    await this.dataSource.manager.transaction(async (entityManager) => {
       const userRepo = entityManager.getRepository(User);
       const userGroupRepo = entityManager.getRepository(UserGroup);
       const userPermissionRepo = entityManager.getRepository(UserPermission);
@@ -371,7 +372,7 @@ export class UserService implements UserServiceInterface {
         'Username should be provided with email or phone',
       );
     }
-    let query = await this.userRepository.getQueryBuilder('user');
+    let query = this.userRepository.createQueryBuilder('user');
     if (email) {
       query = query.orWhere('lower(user.email) = lower(:email)', {
         email: nullCheckedEmail,
