+++ src/database/database.constants.ts
@@ -1,0 +1,1 @@
+export const TENANT_CONNECTION = 'TENANT_CONNECTION';
